Extract `StateReader` into a `State` submodule
